fix: resolve remaining MSVC compilation errors (round 3)
Fixes across 18 files:
- CyberpunkTheme: FontOptions -> Font constructor (String::operator bool)
- EQEffect: static_cast jlimit args to float
- PhysicalModel: add <random> include
- SpectralDNA: ValueTree .isValid() instead of bool
- VoiceManager: clearCurrentNote -> clearNote, base class startVoice
- MacroController: getChildIterator for XmlElement
- GenerativeTimbreDesigner: LUT params for latentToPartialsCore
- MeteringEngine: fix ebur128_state forward decl
- PluginProcessor: add missing SaturationEffectAdapter
- ModulationAssignPanel: -> instead of . on unique_ptr
- XYPad: setTooltip -> setToolTip
- SpectrumEditorCanvas: cast to int for drawRect
- EffectRackComponent: PopupMenu::Item for addItem callback
- PresetManager: include Randomizer/MidiLearn, fix String compare
- VocalProcessor: ProcessContextReplacing for AudioBlock
- SaturationEffect: function member instead of setFunction
- EffectsChain: vector<EffectSlot> -> list<EffectSlot>

+++ src/dsp/effects/EQEffect.h
@@ -51,9 +51,9 @@ public:
             int idx = band.getProperty("index", 0);
             if (idx >= 0 && idx < 3)
             {
-                bands[idx].frequency = juce::jlimit(20.0f, 20000.0f, band.getProperty("frequency", 1000.0f));
-                bands[idx].gain      = juce::jlimit(-24.0f, 24.0f, band.getProperty("gain", 0.0f));
-                bands[idx].q         = juce::jlimit(0.1f, 10.0f, band.getProperty("q", 0.707f));
+                bands[idx].frequency = juce::jlimit(20.0f, 20000.0f, static_cast<float>(band.getProperty("frequency", 1000.0f)));
+                bands[idx].gain      = juce::jlimit(-24.0f, 24.0f, static_cast<float>(band.getProperty("gain", 0.0f)));
+                bands[idx].q         = juce::jlimit(0.1f, 10.0f, static_cast<float>(band.getProperty("q", 0.707f)));
                 bands[idx].type      = static_cast<EQBandType>(juce::jlimit(0, 2, static_cast<int>(band.getProperty("type", static_cast<int>(EQBandType::Peaking)))));
                 updateCoeffs(idx);
             }
